stabilize stroke capture during brush drags

+++ monster.cpp
@@ -6,7 +6,7 @@ using namespace std;
 monster::monster() {}
 
 Stroke monster::makeDummyBody() {
-    std::vector<Eigen::Vector2d> points = {
+    std::vector<Eigen::Vector2f> points = {
                                            {-1.0,  0.0},
                                            {-0.7,  0.7},
                                            { 0.0,  1.0},
@@ -20,7 +20,7 @@ Stroke monster::makeDummyBody() {
 }
 
 Stroke monster::makeDummyLeg() {
-    std::vector<Eigen::Vector2d> points = {
+    std::vector<Eigen::Vector2f> points = {
         {-0.3, -0.7},  // top left, sits on body boundary
         {-0.3, -1.5},  // bottom left
         { 0.3, -1.5},  // bottom right
